requirements.txt
motor
uvloop
aiofiles
dnspython
pyrofork
humanfriendly
python-dotenv
tgcrypto-pyrofork
aiohttp[speedups]
